Plugin State Management Design Fixes
The renderer cannot clear the state of stateful plugins for _every render_ because it defaults the point of maintaining any state.

This is particularly obvious when using the partial plugin, which may very well mutate state in stateful plugins, but because this plugin simply defers to the template renderer, it was clearing the state of all plugins, just by rendering a partial.

The template renderer, now composes a persistent plugin proxy for tracking plugin execution and exposes a method to clear state from the proxy context - which is a better place for this behaviour anyway.

Because the template renderer cannot possibly know _when_ it should clear plugin state, a method is added to the interface to proxy to the, er, proxy.

Actually clearing the state occurs in `View` now, after we have performed a "full" render.

The next logical step is to create an interface for a "Plugin Manager" that replaces the proxy, and use dependency injection to make the plugin manager available to everything else that might need to either use it, or manage plugin state.

Signed-off-by: George Steel <george@net-glue.co.uk>